+++ android/renderscript/ProgramFragmentFixedFunction.java
@@ -17,6 +17,9 @@
 package android.renderscript;
 
 
+import android.util.Log;
+
+
 /**
  * @hide
  * @deprecated in API 16
@@ -28,7 +31,7 @@ package android.renderscript;
  *
  **/
 public class ProgramFragmentFixedFunction extends ProgramFragment {
-    ProgramFragmentFixedFunction(long id, RenderScript rs) {
+    ProgramFragmentFixedFunction(int id, RenderScript rs) {
         super(id, rs);
     }
 
@@ -49,7 +52,7 @@ public class ProgramFragmentFixedFunction extends ProgramFragment {
          */
         public ProgramFragmentFixedFunction create() {
             mRS.validate();
-            long[] tmp = new long[(mInputCount + mOutputCount + mConstantCount + mTextureCount) * 2];
+            int[] tmp = new int[(mInputCount + mOutputCount + mConstantCount + mTextureCount) * 2];
             String[] texNames = new String[mTextureCount];
             int idx = 0;
 
@@ -71,7 +74,7 @@ public class ProgramFragmentFixedFunction extends ProgramFragment {
                 texNames[i] = mTextureNames[i];
             }
 
-            long id = mRS.nProgramFragmentCreate(mShader, texNames, tmp);
+            int id = mRS.nProgramFragmentCreate(mShader, texNames, tmp);
             ProgramFragmentFixedFunction pf = new ProgramFragmentFixedFunction(id, mRS);
             initProgram(pf);
             return pf;
